feat: extracted all routes to a separate file
